Accept specific asset config in a generic asset context, fix include path routing
Because the asset in the compiler is the generic one, the compiler is unable to know specific details about the non-generic asset. Passing these values down the chain and accepting them here allows the compiler to access this asset-specific information in a generic context.

Full paths are now supported by the include parser, instead of the singular file name which was breaking it before. Assets are now also forced to be relative to the document root by removing prefixing slashes.